Made tons of types more explicitly signed/sized
Accel:
- accel_start_record_gesture takes a uint16_t *gesture
- accel_end_record_gesture takes a uint16_t *gesture

Moving_avg:
- wbuf_len is now a uint16_t

Updated pebble sample accordingly

+++ sample/simple-accelerometer/src/moving_avg_ticker.h
@@ -18,7 +18,7 @@ typedef struct moving_avg_values {
     // Circular buffer
     int *wbuf;
     int wbuf_end;
-    int wbuf_len;
+    uint16_t wbuf_len;
 
     int subtotal;
     int subtotal_size;
